Change data type from float16 to float32

+++ financials.py
@@ -9,11 +9,11 @@ def get_madad(initial_value=params.MADAD_INITIAL_VALUE,
               steady_ramp_months_duration=params.STEADY_RAMP_MONTHS_DURATION,
               long_range_centerline=params.LONG_RANGE_CENTERLINE):
 
-    madad_0 = initial_value * np.ones(initial_steady_period_months_duration, dtype='float16')
+    madad_0 = initial_value * np.ones(initial_steady_period_months_duration, dtype='float32')
     madad_1 = np.polyval([(long_range_centerline - initial_value) / steady_ramp_months_duration, madad_0[-1]],
                          np.arange(steady_ramp_months_duration))
     madad_2 = madad_1[-1] * np.ones(params.MAX_DURATION)
-    madad = np.concatenate([madad_0, madad_1, madad_2]).astype('float16')
+    madad = np.concatenate([madad_0, madad_1, madad_2]).astype('float32')
     return madad
 
 
@@ -28,11 +28,11 @@ def get_prime(initial_value=params.PRIME_INITIAL_VALUE,
         initial_value = requests.get('https://Boi.org.il/PublicApi/GetInterest').json()['currentInterest'] / 100
     except:
         pass
-    prime_0 = initial_value * np.ones(initial_steady_period_months_duration, dtype='float16') + banks_margine
+    prime_0 = initial_value * np.ones(initial_steady_period_months_duration, dtype='float32') + banks_margine
     prime_1 = np.polyval([(long_range_centerline - prime_0[-1]) / steady_ramp_months_duration, prime_0[-1]],
                          np.arange(steady_ramp_months_duration))
     prime_2 = np.sin(2 * np.pi * np.arange(0, 30, 1 / 12) / wavewlwngthy) * amplitude + prime_1[-1]
-    prime = np.concatenate([prime_0, prime_1, prime_2]).astype('float16')
+    prime = np.concatenate([prime_0, prime_1, prime_2]).astype('float32')
     return prime
 
 
@@ -53,7 +53,7 @@ def get_spitzer_amortization(monthly_rate_array: np.ndarray, duration: int, prin
     ipmt = - npff.ipmt(monthly_rate_array[:duration], per, duration, principal) # interest portion of a payment
     ppmt = (principal / duration) * np.ones_like(per)  # payment against loan principal
     pmt = ((ipmt.sum() + principal) / duration) * np.ones_like(per) # monthly payment against loan principal plus interest
-    return {'pmt': pmt.astype('float16'), 'ipmt': ipmt.astype('float16'), 'ppmt': ppmt.astype('float16')}
+    return {'pmt': pmt.astype('float32'), 'ipmt': ipmt.astype('float32'), 'ppmt': ppmt.astype('float32')}
 
 
 def get_equal_amortization(monthly_rate_array, duration, principal) -> dict:
@@ -61,7 +61,7 @@ def get_equal_amortization(monthly_rate_array, duration, principal) -> dict:
     ppmt = principal / duration * np.ones_like(per) # payment against loan principal
     ipmt = (principal - ppmt.cumsum()) * monthly_rate_array[:duration] # interest portion of a payment
     pmt = ipmt + ppmt # monthly payment against loan principal plus interest
-    return {'pmt': pmt.astype('float16'), 'ipmt': ipmt.astype('float16'), 'ppmt': ppmt.astype('float16')}
+    return {'pmt': pmt.astype('float32'), 'ipmt': ipmt.astype('float32'), 'ppmt': ppmt.astype('float32')}
 
 
 fixed_yearly_risk_rate = lambda r: interp1d([0., .45, .6, .7, .75],
@@ -88,9 +88,9 @@ def update_yearly_to_monthly_rates_with_risk(funding_rate: float, is_married_cou
     madad_rate = monthly_changing_yearly_MADAD * risk['madad_yearly_added_risk_yearly_rate']
     prime_rate = (monthly_changing_yearly_PRIME + params.PRIME_ADDED_YEARLY_RATE) * risk['fixed_yearly_added_risk_rate']
     prime_rate[prime_rate < 0.0005] = 0.0005
-    return {'fixed_rate': yearly_rate_to_monthly(fixed_rate).astype('float16'),
-            'madad_rate': yearly_rate_to_monthly(madad_rate).astype('float16'),
-            'prime_rate': yearly_rate_to_monthly(prime_rate).astype('float16')}
+    return {'fixed_rate': yearly_rate_to_monthly(fixed_rate).astype('float32'),
+            'madad_rate': yearly_rate_to_monthly(madad_rate).astype('float32'),
+            'prime_rate': yearly_rate_to_monthly(prime_rate).astype('float32')}
